+++ templates/templates/report/VerificationReportTemplate.html
@@ -177,6 +177,7 @@
                 <!-- COMMENT: Again, we would need to get this after the merge of the main->release branch PR, which means we need to be able to find that specific PR -->
                 <!-- list all enabled URS -->
                 <p><var>LIST_OF_REQUIREMENTS</var></p>
+                <p><var>REQUIREMENT_CONTENTS_AND_APPROVERS</var></p>
             </article>
         
             <!-- list markdown files in the dedicated design folder at the time of pull request completion -->
